added randomizer and locking stratagems. Gonna work on getting the lock to disable the button
